Update hex order parameter to use floats

+++ cpp/order/HexOrderParameter.h
@@ -33,7 +33,7 @@ class HexOrderParameter
         void computePy(boost::python::numeric::array points);
                        
         //! Get a reference to the last computed psi
-        boost::shared_array< std::complex<double> > getPsi()
+        boost::shared_array< std::complex<float> > getPsi()
             {
             return m_psi_array;
             }
@@ -41,7 +41,7 @@ class HexOrderParameter
         //! Python wrapper for getPsi() (returns a copy)
         boost::python::numeric::array getPsiPy()
             {
-            std::complex<double> *arr = m_psi_array.get();
+            std::complex<float> *arr = m_psi_array.get();
             return num_util::makeNum(arr, m_Np);
             }
 
@@ -51,7 +51,7 @@ class HexOrderParameter
         locality::LinkCell m_lc;          //!< LinkCell to bin particles for the computation
         unsigned int m_Np;                //!< Last number of points computed
         
-        boost::shared_array< std::complex<double> > m_psi_array;         //!< psi array computed
+        boost::shared_array< std::complex<float> > m_psi_array;         //!< psi array computed
     };
 
 //! Exports all classes in this file to python
